+++ spec/models/activity_spec.rb
@@ -1233,12 +1233,14 @@ RSpec.describe Activity, type: :model do
   end
 
   describe "#actual_total_for_report_financial_quarter" do
+    let(:current_quarter) { FinancialQuarter.for_date(Date.today) }
+
     it "returns the total of all the activity's transactions scoped to a report" do
       project = create(:project_activity, :with_report)
       report = Report.for_activity(project).first
-      create(:transaction, parent_activity: project, value: 100.20, report: report, date: Date.today)
-      create(:transaction, parent_activity: project, value: 50.00, report: report, date: Date.today)
-      create(:transaction, parent_activity: project, value: 210, report: report, date: 4.months.ago)
+      create(:transaction, parent_activity: project, value: 100.20, report: report, **current_quarter)
+      create(:transaction, parent_activity: project, value: 50.00, report: report, **current_quarter)
+      create(:transaction, parent_activity: project, value: 210, report: report, **current_quarter.pred)
 
       expect(project.actual_total_for_report_financial_quarter(report: report)).to eq(150.20)
     end
@@ -1246,8 +1248,8 @@ RSpec.describe Activity, type: :model do
     it "does not include the totals for any transactions outside the report's date range" do
       project = create(:project_activity, :with_report)
       report = Report.for_activity(project).first
-      create(:transaction, parent_activity: project, value: 100.20, report: report, date: 6.months.ago)
-      create(:transaction, parent_activity: project, value: 210, report: report, date: 4.months.ago)
+      create(:transaction, parent_activity: project, value: 100.20, report: report, **current_quarter.pred.pred)
+      create(:transaction, parent_activity: project, value: 210, report: report, **current_quarter.pred)
 
       expect(project.actual_total_for_report_financial_quarter(report: report)).to eq(0)
     end
@@ -1257,7 +1259,7 @@ RSpec.describe Activity, type: :model do
     let(:project) { create(:project_activity) }
 
     it "returns the activity's planned disbursement value for a report's financial quarter" do
-      forecast = PlannedDisbursementHistory.new(project, 3, 2020)
+      forecast = PlannedDisbursementHistory.new(project, financial_quarter: 3, financial_year: 2020)
       reporting_cycle = ReportingCycle.new(project, 2, 2020)
 
       reporting_cycle.tick
@@ -1270,8 +1272,8 @@ RSpec.describe Activity, type: :model do
     end
 
     it "does not include totals for any planned disbursements outside the report's date range" do
-      q3_forecast = PlannedDisbursementHistory.new(project, 3, 2020)
-      q4_forecast = PlannedDisbursementHistory.new(project, 4, 2020)
+      q3_forecast = PlannedDisbursementHistory.new(project, financial_quarter: 3, financial_year: 2020)
+      q4_forecast = PlannedDisbursementHistory.new(project, financial_quarter: 4, financial_year: 2020)
       reporting_cycle = ReportingCycle.new(project, 2, 2020)
 
       reporting_cycle.tick
@@ -1285,7 +1287,7 @@ RSpec.describe Activity, type: :model do
     end
 
     it "only counts the latest revision value for a planned disbursement" do
-      forecast = PlannedDisbursementHistory.new(project, 3, 2020)
+      forecast = PlannedDisbursementHistory.new(project, financial_quarter: 3, financial_year: 2020)
       reporting_cycle = ReportingCycle.new(project, 1, 2020)
 
       reporting_cycle.tick
@@ -1311,7 +1313,7 @@ RSpec.describe Activity, type: :model do
 
     let(:project) { create(:project_activity) }
     let(:reporting_cycle) { ReportingCycle.new(project, 2, 2020) }
-    let(:forecast) { PlannedDisbursementHistory.new(project, 3, 2020) }
+    let(:forecast) { PlannedDisbursementHistory.new(project, financial_quarter: 3, financial_year: 2020) }
 
     it "returns the variance between #actual_total_for_report_financial_quarter and #forecasted_total_for_report_financial_quarter" do
       reporting_cycle.tick